@@ -1,8 +1,8 @@
 package com.u1fukui.bbs.helper
 
 import android.content.Context
-import android.support.annotation.StringRes
-import android.support.v7.app.AlertDialog
+import androidx.annotation.StringRes
+import androidx.appcompat.app.AlertDialog
 
 import com.u1fukui.bbs.R
 
