@@ -2,7 +2,7 @@ package com.u1fukui.bbs.utils
 
 
 import android.content.Context
-import android.support.annotation.StringRes
+import androidx.annotation.StringRes
 import android.widget.Toast
 import com.u1fukui.bbs.network.ApiErrorMesageResolver
 
