@@ -1,9 +1,9 @@
 package com.u1fukui.bbs.customview
 
 import android.content.Context
-import android.databinding.BindingAdapter
-import android.databinding.DataBindingUtil
-import android.databinding.ViewDataBinding
+import androidx.databinding.BindingAdapter
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
 import android.util.AttributeSet
 import android.view.Gravity
 import android.view.LayoutInflater
