@@ -1,8 +1,8 @@
 package com.u1fukui.bbs.helper
 
 
-import android.databinding.ObservableField
-import android.databinding.ObservableInt
+import androidx.databinding.ObservableField
+import androidx.databinding.ObservableInt
 import android.view.View
 import com.u1fukui.bbs.network.ApiErrorMesageResolver
 
