@@ -1,7 +1,7 @@
 package com.u1fukui.bbs.paging.comment
 
-import android.arch.lifecycle.MutableLiveData
-import android.arch.paging.PageKeyedDataSource
+import androidx.lifecycle.MutableLiveData
+import androidx.paging.PageKeyedDataSource
 import com.u1fukui.bbs.model.Comment
 import com.u1fukui.bbs.paging.NetworkState
 import com.u1fukui.bbs.repository.ThreadRepository
