@@ -1,7 +1,7 @@
 package com.u1fukui.bbs.ui.creation.thread
 
 
-import android.arch.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProviders
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
